Split Form Section Component into several child components
