+++ src/components/body-editor/body-editor.module.scss
@@ -1,0 +1,18 @@
+.bodyEditor {
+    box-sizing: border-box;
+    border: 1px solid #ddd;
+    cursor: text;
+    padding: 16px;
+    border-radius: 2px;
+    margin-bottom: 2em;
+    box-shadow: inset 0px 1px 8px -3px #ABABAB;
+    background: #fefefe;
+}
+
+.bodyEditor :global(.public-DraftEditor-content) {
+    min-height: 140px;
+}
+
+.allTools {
+    display: flex;
+}
