+++ src/components/common-title/common-title.styles.scss
@@ -6,5 +6,4 @@
     font-size: 1rem;
     font-weight: 400;
     height: 50%;
-    // border: 1px solid red;
 }
