+++ src/pages/create-article-page/create-article-page.styles.scss
@@ -1,5 +1,5 @@
 .create-article-page-container {
-    border: 1px solid yellow;
+    // border: 1px solid yellow;
     display: flex;
     justify-content: center;
     min-height: 100vh;
@@ -8,7 +8,7 @@
 .create-article-page-wrapper {
     display: grid;
     grid-template-rows: 0.3fr 1fr 2.5fr;
-    border: 1px solid blue;
+    // border: 1px solid blue;
     width: 80%;
     padding-inline: 1.5rem;
     margin-block: 5%;
